pkg/lib: move event/fsinfo/_internal into cockpit/
Having these in the "global" namespace alongside everything in
node_modules seems like a disaster waiting to happen.  Let's make a
`cockpit` dir and import them like `'cockpit/fsinfo'`, etc.
