+++ pkg/lib/cockpit/_internal/base64.js
