+++ pkg/lib/cockpit/_internal/channel.js
