+++ pkg/lib/cockpit/_internal/common.ts
