+++ pkg/lib/cockpit/_internal/deferred.js
