+++ pkg/lib/cockpit/_internal/location.ts
