+++ pkg/lib/cockpit/_internal/parentwebsocket.js
