+++ pkg/lib/cockpit/event.ts
