+++ pkg/lib/cockpit/fsinfo.ts
